init: ensure initcall ordering with LTO
With LTO, the compiler doesn't necessarily obey link order for
initcalls, and the initcall variable needs to be globally unique
to avoid naming collisions.

In order to preserve the correct order, we add each variable
into its own section and generate a linker script (in
scripts/link-vmlinux.sh) to ensure the order remains correct.  We
also add a __COUNTER__ prefix to the name, so we can retain the
order of initcalls within each compilation unit, and __LINE__ to
make the names more unique.

Change-Id: I6828c9140d328c5155abff6337de1d34e789201e
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>
[wight554: backport to 4.4]
Signed-off-by: Volodymyr Zhdanov <wight554@gmail.com>