kbuild: do not drop -I without parameter
The comment line for addtree says "skip if -I has no parameter".

What it actually does is "drop if -I has no parameter".  For example,
if you have the compiler flag '-I foo' (a space between), it will be
converted to 'foo'.  This completely changes the meaning.

What we want is, "do nothing" for -I without parameter so that
'-I foo' is kept as-is.

Change-Id: I09d8c7461250e956f0fd2ed2799e61667b2509c8
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>